move sidx requesting into the dash playlist loader
